Move all public types to public-types.ts
