+++ wasm-node/javascript/src/index-deno.ts
@@ -15,7 +15,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { Client, ClientOptions, start as innerStart } from './client.js'
+import { Client, ClientOptions } from './public-types.js'
+import { start as innerStart } from './client.js'
 import { Connection, ConnectionConfig } from './instance/instance.js';
 import { default as wasmBase64 } from './instance/autogen/wasm.js';
 
@@ -31,7 +32,7 @@ export {
     QueueFullError,
     JsonRpcDisabledError,
     LogCallback
-} from './client.js';
+} from './public-types.js';
 
 /**
  * Initializes a new client. This is a pre-requisite to connecting to a blockchain.
